+++ multiqc/modules/htseq/htseq.py
@@ -86,11 +86,11 @@ class MultiqcModule(BaseMultiqcModule):
             'format': '{:.1f}%'
         }
         headers['assigned'] = {
-            'title': 'M Assigned',
-            'description': 'Assigned reads (millions)',
+            'title': '{} Assigned'.format(config.read_count_prefix),
+            'description': 'Assigned Reads ({})'.format(config.read_count_desc),
             'min': 0,
             'scale': 'PuBu',
-            'modify': lambda x: float(x) / 1000000,
+            'modify': lambda x: float(x) * config.read_count_multiplier,
             'shared_key': 'read_count'
         }
         self.general_stats_addcols(self.htseq_data, headers)